add review-bot to require fellows as reviewers (#31)
* add review-bot to require fellows as reviewers

Created a Github Action that uses the [Review-Bot app](https://github.com/paritytech/review-bot) to require fellows to review pull requests before allowing the PR to be merged.

The user's information is fetched always from the chain after every event.

It looks in the fellows data for a field named GitHub and it extracts the handle from there. This resolves #7 (you can find more information about the request there)

This uses [`pull_request_target`](https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#pull_request_target) for the event, not `pull_request`. This is a security measure so that an attacker doesn’t have access to the secrets.

* removed non existent directory

* Apply suggestions for required rankings

Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: joe petrowski <25483142+joepetrowski@users.noreply.github.com>

* merged system parachains with relay files

* updated minFellowsRank for .github files

Co-authored-by: Bastian Köcher <git@kchr.de>

* updated file to new version

Version 2.0.0 requires new types and fields.

This will help us so we can develop custom rules in the future.

* added CHANGELOG to 'Relay and System files' rule

---------

Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: joe petrowski <25483142+joepetrowski@users.noreply.github.com>