Added GitHub action to auto-merge (#55)
* added github action to enable auto-merge in PRs

Added two GitHub actions to enable auto-merge in the repository.

This resolves #41

## [Auto-Merge-Bot](https://github.com/paritytech/auto-merge-bot)

This bot allows _public_ members of the organization and the author of the PR to enable/disable auto-merge in a PR. It also has a field for a list of users who are also allowed to trigger the bot. For that we have the second action.

* added step to allow fellows to trigger the bot

## [Get Fellows Action](https://github.com/paritytech/get-fellows-action)

This action gets the fellows from the chain data and it extracts their GitHub handles. This handles are then input into the `auto-merge-bot` action as a list of users who are allowed to trigger the bot.

* added a pull request template

I have also updated the Pull Request template to mention the existence of this bot to new contributors.

* added mention of the bot in the readme

Updated the readme to contain information about the bot, how to use it, and a link to its documentation