Improve release process
This pull request introduces the `changelog-processor.py`. A small
script for getting some information out of the `CHANGELOG.md` file for
improving the release process. The script supports to get the latest
version in the `CHANGELOG.md` file, aka the version top of the
file. (This could also be `Unreleased`) The other command is for
printing if there should be done a release or not. This is decided based
on the latest version and if there exists a tag for it.