Improve release process (#63)
* Improve release process

This pull request introduces the `changelog-processor.py`. A small
script for getting some information out of the `CHANGELOG.md` file for
improving the release process. The script supports to get the latest
version in the `CHANGELOG.md` file, aka the version top of the
file. (This could also be `Unreleased`) The other command is for
printing if there should be done a release or not. This is decided based
on the latest version and if there exists a tag for it.

* Use correct syntax

* Single quotes..

* Integer?

* :facepalm:

* This?

* AHH

* Some last improvements

* Update .github/changelog-processor.py

Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>

* Apply suggestions from code review

* Make the arguments mutally exclusive

---------

Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>