Restore governance lock periods to 7 days in Polkadot (#86)
This is a revert of
https://github.com/paritytech/polkadot/commit/d73503cc2351ffec047a4912ecdb3e1eab5f46aa
which has not been released.

Deploying without this change would lead to referendum participants and
delegates having their tokens locked for 4 times longer than expected.

Participants in any referendums ending after the runtime upgrade is
applied would be affected.
Even more of an issue is delegates, where all current delegates would be
affected by 4 times longer lock up issues. The affected user-base is
enormous, and there is little to no possibility of reaching all these
users to get them to adapt in time.

Although a 7 Day lock multiplier wasn't the originally intended target
for Polkadot, breaking user expectations with this change is an even
larger issue.

---------

Co-authored-by: Bastian Köcher <info@kchr.de>
Co-authored-by: Bastian Köcher <git@kchr.de>