use github app for auto-merge-bot
This changes the credentials for auto-merge-bot

With this it stops using a GitHub action but instead it uses a GitHub app, by using a GitHub app, PRs merged by it will trigger GitHub actions.

This is to circumvent a security measure set up by GitHub to stop recursive action executions.