feat: align `LocationToAccountId` derivation (#67)
In order to allow other chains to have consistent "remote" accounts in
other chains, it is needed that all chains in the ecosystem derive
accounts in the XCM logic in the same way.

See
https://forum.polkadot.network/t/multichain-friendly-account-abstraction/1298/23

This PR aligns all runtimes of system-parachains to have that
derivation. It is debatable whether this derivation should also go into
the relay-chain runtimes.