Lower System Parachain Existential Deposits (#131)
As referenced in a few discussions (e.g.
[forum](https://forum.polkadot.network/t/lowering-or-removing-existential-deposit-in-the-long-run/2167/10)),
the Asset Hub Existential Deposit was set at genesis to 1/10th of the
Relay Chain's. The goal however was always to lower it as the Relay
Chain became more efficient at executing parachains to reflect the lower
resource consumption on parachains and provide stronger incentives to
prefer parachains for user-applications over the Relay Chain.

This PR lowers the system chain existential deposits by a factor of ten,
to 1/100th of the Relay Chain's ED, or 0.01 DOT.

For discussion/comment: Should all system paras have a lower ED, or just
Asset Hub? On one hand, it is less confusing if they are all the same.
On the other, we actually _don't_ necessarily want to encourage lots of
accounts on a chain like Bridge Hub. For this PR, I've adjusted all of
the system paras, but in separate commits.

---------

Co-authored-by: Bastian Köcher <info@kchr.de>