Add kusama and polkadot SP constants (#115)
This PR brings the Kusama and Polkadot constants for System Parachains
into the control of the Fellowship.

Following up on https://github.com/paritytech/polkadot-sdk/pull/2620 and
paired with https://github.com/paritytech/polkadot-sdk/pull/2666.

Types are still used from the `parachains-common` crate, but constants
are moved into a new crate `system-parachains-constants` in this repo.

The re-exports `pub use parachains_common as common` in both asset hubs
and collectives have been removed. This is a breaking change for anybody
using them in this way, but these should be directly imported from
`parachains-common`. This should be noted on the changelog.

---------

Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: Bastian Köcher <git@kchr.de>