added up-to-date-action
Added action to keep pull requests up to date. Every time master receives a push, it will look for all the PRs and sync them with the master branch.

This only works when the PR has `auto-merge` enabled (it can be enabled for all PRs too if required).

Merging this PR will count as solving paritytech/auto-merge-bot#21 and paritytech/up-to-date-action#1

We use a github app to solve the issue that tests would not run if the credentials comes from a github action.