added action to verify the changelog is updated
Added a GitHub action that uses foodee/pr-includes-file-change to fetch all the changes and verify that the file `CHANGELOG.md` has been modified.

If it has not been modified and the PR body does not include the text `[x] Does not require a CHANGELOG entry`, the action will fail.

This will fulfill the requirements for polkadot-fellows/runtimes#143