added action to verify the changelog is updated (#153)
Added a GitHub action that uses
[foodee/pr-includes-file-change](https://github.com/marketplace/actions/pr-includes-file-change)
to fetch all the changes and verify that the file `CHANGELOG.md` has
been modified.

If it has not been modified and the PR body does not include the text
`[x] Does not require a CHANGELOG entry`, the action will fail.

This resolves polkadot-fellows/runtimes#143

The template was updated to include a reminder on how to modify the
changelog and instructions on how to skip the CI if the PR does not
require to modify the changelog

---------

Co-authored-by: fellowship-merge-bot[bot] <151052383+fellowship-merge-bot[bot]@users.noreply.github.com>