set events for changelog action to work on edit (#167)
Enabled events `synchronize` and `edited` for the [changelog
action](https://github.com/polkadot-fellows/runtimes/blob/main/.github/workflows/changelog.yml).

`synchronize` works when the branch is modified (commit is pushed).

`edited` works when the description of the PR is modified.

`edited` is not enabled by default, so we had to add it.

- [x] Does not require a CHANGELOG entry