Review Bot version 2.4.0 - Fellow Scores (#168)
Updated Review Bot to version 2.4.0 which contains
paritytech/review-bot#110 (solved in paritytech/review-bot#113)

This resolves #93 

Updated review bot to latest version which brings new changes to the
fellow object:

Added new optional field to the config named `scores` that has the score
of each fellow per dan (added up to dan IX).

When the rule of type `fellows` has the optional field of
`minTotalScore` it will check that the score of the fellows who approved
the PR sums up to that number. If it doesn't sum to that number it will
fail with a custom error listing the score of each fellow for reference
(and saying how many points are missing).

Find a complete explanation of the new configuration here: [Fellows
Rule](https://github.com/paritytech/review-bot/blob/v2.4.0/README.md#fellows-rule)

- [x] Does not require a CHANGELOG entry

I also added some scores to all the dans (feel free to modify them) and
updated the `Relay and system files` rule to have a minimum score of 6.