Deprecate `xcm::body::TREASURER_INDEX` constant, use standard `Treasury` varian instead (#149)
Deprecate the `xcm::body::TREASURER_INDEX` constant and use the standard
`Treasury` variant from the `xcm::BodyId` type instead.

The custom body index is redundant here since the standard `Treasury`
body variant exists. It was overlooked previously, and the index was
added by mistake.
Removing the index should not affect anything. It's only used as one of
the origins that guards permissioned operations for the fellowship
treasury instance and asset rate pallet instance, and not used to derive
any account ids.