Add a "Wish for Change" Track (#184)
Posting this PR for discussion, but perhaps people have other solutions
in mind.

The Root track was designed to handle only one referendum at a time,
because calls requiring Root are usually sensitive and should be
evaluated one at a time. However, we have also used the Root track to
make remarks to signal desires/wishes of the network to various bodies
in the network (e.g. RFC-12, instructing the Fellowship to add a
collective).

These statements do not execute any stateful logic that would affect the
network, and in my opinion more than one could be evaluated at a time.
These referenda should not hold up voting on proposals that actually do
require Root, nor be forced to queue (for example, it should be possible
to propose two new collectives in parallel).

The approval/support requirements are the same as Root, but the origin
does not map to any privilege. Passing something on the track is merely
a signal.

---------

Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>