Fix default bridge byte fee to use TransactionByteFee of the bridged chain (#174)
This PR does two things:
1) (tiny refactoring) adds `TRANSACTION_BYTE_FEE` constant to
`system-parachains-constants` crate. We are using the same value (`1 *
MILLICENTS`) across all chains, so it makes sense to move it there;
2) changes default value of governance-controlled
`XcmBridgeHubRouterByteFee` constant (which is used to cover the cost of
message delivery within the remote consensus) to use
`TRANSACTION_BYTE_FEE` of the bridged chain (also does 1:5/5:1
conversion). Since governance has not set value of this constant yet, we
are changing the default value here, in the code.

~The failing test is fixed by the #137, so it makes sense to merge this
PR only after #137~ Updated

---------

Co-authored-by: Branislav Kontur <bkontur@gmail.com>
Co-authored-by: Kian Paimani <5588131+kianenigma@users.noreply.github.com>