updated auto-merge-bot to 1.0.1
This upgrade contains the change in paritytech/auto-merge-bot#25 which fixes the problem with unstable branches.

## The problem

When the `/merge` command is triggered on a PR that still can't be merge, the bot enables auto-merge.

**But**, when the command is run when the PR is ready to merge, the action will fail to enable auto-merge (this is expected).

We evaluate the error message, and if it says that the PR is clean, it simply merges the PR.

But, we did not handle `unstable` PRs.

An unstable PR is a PR that _can be merged, but a **non required status check is failing**._

So, if a user invoked the `/merge` command on a PR that was ready to merge, but a not-required status check was failing, it would crash.

This upgrade fixes this problem.