[async_backing] Bump ParachainHost to api version 10 on polkadot (#222)
This will enable async-backing subsystems on polkadot, the relaychain
would work in legacy mode where only candiddates built on top of
pervious relay chain are activated.

## Notes
- Had to bring an unrelated change as well `minimum_backing_votes`
because that was the v6, we can't skip versions, the value of that
configuration on polkadot is 2, so that's what is going to be used after
this runtime update. Changes have been running on kusama.
- `disabled_validators`, `node_features`, `approval_voting_params` are
not related with async-backing, but given they are low risk, we decided
to include them as well, see comments.

## Known risks

Async backing subsytems is a major change in the way collator-protocol
and the backing work, so there are still some unknowns if this is
completely bug free. It has been running on kusama for a month already,
but not without issues:
- Validators that did not upgrade to compatible versions will not be
able to participate in backing, so if enough of those are randomly
picked that group won't be able to back anything. With
backing_group_size = 5 and minimum_backing_votes = 2, 10% validator not
upgraded, that chance is about 2.5%.

- Additionally, same un-upgraded groups won't be able to include the
backing votes on chain when they author blocks, so 10% of the blocks
won't back any candidates.

- We are still not sure if item 2) from here
https://github.com/paritytech/polkadot-sdk/issues/3314#issuecomment-1953542045
is caused by async backing, the observable issue is sometimes after
restart/upgrade some validators are getting 0 rewards and apparently
they are not backing anything.


<!-- Remember that you can run `/merge` to enable auto-merge in the PR
-->

<!-- Remember to modify the changelog. If you don't need to modify it,
you can check the following box.
Instead, if you have already modified it, simply delete the following
line. -->

---------

Signed-off-by: Alexandru Gheorghe <alexandru.gheorghe@parity.io>
Co-authored-by: fellowship-merge-bot[bot] <151052383+fellowship-merge-bot[bot]@users.noreply.github.com>