Keep `im-online` events decodable after pallet removal (#235)
This is a follow-up to #178, which aims to keep the `im-online` pallet
events decodable after the pallet itself is removed.

As
[discussed](https://github.com/polkadot-fellows/runtimes/pull/178#issuecomment-1977608363),
the inability to decode events may result in missing a candidate in
approval voting and a parachain finality stall. Although the
circumstances of such an event are unlikely, it's better to safeguard
them.

The transient code introduced in this PR should be removed after the
upgrade is enacted.

- [x] Does not require a CHANGELOG entry