Remove state trie migration from kusama, add it to polkadot (#170)
This PR add the state trie migration to V1 to polkadot runtime.
The configuration is the same as the one that was run on kusama.
As in kusama this PR changes the state-trie-version to v1, so the
migration declared in unreleased Have to run with the next upgrade. In
case the migration is not include, the line `state_version: 1` have to
be roll back to `state_version: 0` as warpsync will stop working until
the end of the migration.

<!-- Remember that you can run `/merge` to enable auto-merge in the PR
-->

<!-- Remember to modify the changelog. If you don't need to modify it,
you can check the following box.
Instead, if you have already modified it, simply delete the following
line. -->

- [ ] Does not require a CHANGELOG entry

---------

Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: Bastian Köcher <info@kchr.de>