Treasury Spend Detects Relative Locations of Native Asset (#233)
Treasury Spend detects the relative locations of native asset to compare
origin's spend limits.

When a spend is approved for a native asset located on Asset Hub, the
AssetRate pallet instance does not recognize the location (parent: 1,
here) as a native asset. Therefore, it requires the asset rate to be set
in advance to the rate of `1`. To address this issue in this PR, we
introduce new types to wrap the `AssetRate` implementation of balance
conversion. These types match the native asset's relative locations and
result in a 1:1 balance conversion in such cases.

In addition we provide integration tests to cover these cases.

Examples:
Treasury pallet on Relay Chain accepts the spend of the native asset on
Asset Hub with it's ID presented as `VersionedLocatableAsset` type
```
{
  `location`: (0, Parachain(1000)),
  `asset_id`: (1, Here),
}
```
and it's balance should be converted 1:1 by treasury pallet's
`BalanceConverter` type.

Note: I plan to introduce a similar but configurable implementation in
the polkadot-sdk and utilize it here. However, to get this change in the
next release, I open this PR first.

Frame version: https://github.com/paritytech/polkadot-sdk/pull/3659

---------

Co-authored-by: joe petrowski <25483142+joepetrowski@users.noreply.github.com>
Co-authored-by: Branislav Kontur <bkontur@gmail.com>
Co-authored-by: Bastian Köcher <git@kchr.de>