[coretime-kusama] Bump broker and add migration to import leases to fix launch issues (#276)
For the transition to coretime, the relay runtime upgrade includes a
migration which sends a few XCMs to the coretime chain with Transacts.
This calls `set_lease` (a broker extrinsic on the coretime chain) for
all the leases that are currently in the relay state to get them a lease
on the coretime chain.
This happened yesterday but the XCM didn't have enough weight and so the
leases weren't migrated over. We decided to make a runtime upgrade since
there was a bug with the renewals for short leases anyway. This PR
solves both issues.

The bump to the broker pallet fixes a bug where very short leases
(ending in the first 28 days after running start_sales) could not be
renewed.

This PR includes a migration which does three things:
1. Imports a hardcoded list of leases into the state
2. Adds three cores to ensure there are enough cores for everyone to
renew while also leaving the intended 3 for the open market
3. restarts sales

The migration also tests the renewal fix with try-runtime.

- [ ] Does not require a CHANGELOG entry

---------

Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: Bastian Köcher <info@kchr.de>