Add migrations to onboard people immediately and tidy up previous intervention (#286)
When calling the reserve extrinsic after sales have started, the
assignment will be reserved, but two sale period boundaries must pass
before the core is actually assigned.

Since this can take between 28 and 56 days on production networks, a new
extrinsic is introduced to shorten the timeline at the "cost" of more
weight in https://github.com/paritytech/polkadot-sdk/pull/4273.

This essentially performs four actions:
1. Add an additional core for the new reservation
2. Reserve it (applies after two sale boundaries)
3. Add it to the Workplan for the next sale period
4. Add it to the Workplan for the rest of the current sale period

To allow a quick People Chain onboarding on Kusama, a migration is
introduced here which does the same thing without needing to wait for a
release and to upgrade the runtimes repo.

Another migration is added to clean up an outdated assignment in state
from the old sale start before the leases were added. This avoids the
first lease (parachain 2000) losing its core to the new pool core a few
days before it is given one again with the new timeline at the end of
the new sale period 0.

---------

Co-authored-by: joe petrowski <25483142+joepetrowski@users.noreply.github.com>