Release Kusama Coretime 1.2.3 (#293)
Cherry pick #286 onto the 1.2.2 release for a clean 1.2.3 release.

This release is the one proposed in
https://github.com/polkadot-fellows/runtimes/issues/288

The changelog from main is clipped to only include the relevant changes
from the cherry picked commit.
- A migration is added to onboard the people chain without the usual two
sale boundary delay that came with the shift to coretime.
- Another migration is added to clean up an outdated assignment in state
from the old sale start before the leases were added.

More info in the changelog of this PR and PR #286.

- [ ] Does not require a CHANGELOG entry

---------

Co-authored-by: joe petrowski <25483142+joepetrowski@users.noreply.github.com>