Add P<>K bridge manual zombienet test for asset transfer (#198)
<!-- Remember that you can run `/merge` to enable auto-merge in the PR
-->

<!-- Remember to modify the changelog. If you don't need to modify it,
you can check the following box.
Instead, if you have already modified it, simply delete the following
line. -->

- [x] Does not require a CHANGELOG entry

This PR adds a P<>K bridge zombienet test which checks the following:
- transfering DOT from AssetHub Polkadot to AssetHub Kusama works
- transfering wDOT from AssetHub Kusama to AssetHub Polkadot works
- transfering KSM from AssetHub Kusama to AssetHub Polkadot works
- transfering wKSM from AssetHub Polkadot to AssetHub Kusama works

The test is meant to be ran manually as per the instructions in the
`README.md` and is based on the naive bridge testing "orchestrator"
defined in the `polkadot-sdk` repo:
https://github.com/paritytech/polkadot-sdk/tree/master/bridges/testing .
The "orchestrator" is downloaded using `git sparse checkout` directly
from the `polkadot-sdk` repo (I couldn't find a better way to package
the `bash` and `js` files on which the test depends).

Proposing to add the test in this repo as per the feedback from the
bridges team and since this way it would be more convenient to run it
before a release or when performing a change to the bridge hubs for
example. If there are concerns about this, the second option would be to
add it to the `parity-bridges-common` repo.

Closes: https://github.com/paritytech/polkadot-sdk/issues/3400