Allow Sending XCM messages using a Signed origin on Kusama (#290)
Since the introduction of this origin control, the XCMv3 release has
been integrated on Kusama, and now the `xcm-pallet` is perfectly able to
prepend a `DescendOrigin` instruction to the sent message when calling
`send`. Allowing signed origins to send messages through the network
would extend the possibilities for better integrations with parachains.

---------

Co-authored-by: fellowship-merge-bot[bot] <151052383+fellowship-merge-bot[bot]@users.noreply.github.com>