Polkadot People Phase 2 (Triggers Release 1.2.7) (#350)
Phase 2 of the Polkadot People launch, as described in the [forum
post](https://forum.polkadot.network/t/people-chain-launch-and-identity-migration-plan/5930).

Should not be applied until the People Chain is running and deposits
have been migrated from the Relay Chain.

Built on top of https://github.com/polkadot-fellows/runtimes/pull/349
and should be merged after.