updated review-bot dependencies to latest
This commit updates get-fellows-action to v1.1.4 where `papi` was updated to latest (`0.9.0`) which contains a fix for polkadot-api/polkadot-api#327. This will accelerate the process of the triggering of review-bot.

Review-bot has also been updated to latest. It has the pull request number being injected, so it will always fetch the latest status when it runs.