updated review-bot dependencies to latest (`2.5.0`) (#373)
This commit updates get-fellows-action to v1.1.4 where `papi` was
updated to latest (`0.9.0`) which contains a fix for
polkadot-api/polkadot-api#327. This will accelerate the process of the
triggering of review-bot.

Review-bot has also been updated to latest. It has the pull request
number being injected, so it will always fetch the latest status when it
runs.

With the fix provided in #369 this should make the whole process work as
intended.

- [x] Does not require a CHANGELOG entry

---------

Co-authored-by: fellowship-merge-bot[bot] <151052383+fellowship-merge-bot[bot]@users.noreply.github.com>