Reduce Bounty Payout Delay (#386)
The bounty payout delay originally existed to provide time for the
Council to remove curators of a bounty if they were not acting in
accordance with its existence. The Council motion time was 7 days, and
thus the bounty delay was 8 days to give the Council time to remove
curators and void payouts prior to a bounty being claimable.

With OpenGov tracks being 28 days, I don't see this as anything more
than a PITA for curators/awardees. Does anyone have motivation for this
delay to be in effect in the current setup? (cc @muharem , @shawntabrizi
)