Improve test names (#399)
The name of the "All changelog tests passing" was clashing with the name
of the "All tests passing" test.

Give the check features test a name.

Both changes are required to mark these tests as required in github.

<!-- Remember that you can run `/merge` to enable auto-merge in the PR
-->

<!-- Remember to modify the changelog. If you don't need to modify it,
you can check the following box.
Instead, if you have already modified it, simply delete the following
line. -->

- [x] Does not require a CHANGELOG entry