Reduce the base fee on Polkadot System Chains (#398)
This PR reduces the base fee on all Polkadot system chains by half. As a
result, the base fee will be roughly 1/20 of the base fee on the
Polkadot relay chain.

This will make all transactions on (non-congested) system chains cheaper
and stimulate usage.

Among others, the change will push the transaction fees for transfers on
Asset Hub below $0.01 which is an important psychological threshold.

An initiative by ecosystem agents.

---------

Co-authored-by: Adrian Catangiu <adrian@parity.io>
Co-authored-by: joe petrowski <25483142+joepetrowski@users.noreply.github.com>